Add pending state to monitors and config screens
