+++ Client/src/Components/Label/index.jsx
@@ -139,10 +139,10 @@ const StatusLabel = ({ status, text, customStyles }) => {
       bgColor: theme.palette.error.bg,
       borderColor: theme.palette.error.light,
     },
-    unknown: {
-      dotColor: theme.palette.unresolved.main,
-      bgColor: theme.palette.unresolved.bg,
-      borderColor: theme.palette.unresolved.light,
+    pending: {
+      dotColor: theme.palette.warning.main,
+      bgColor: theme.palette.warning.bg,
+      borderColor: theme.palette.warning.light,
     },
     "cannot resolve": {
       dotColor: theme.palette.unresolved.main,
@@ -176,7 +176,7 @@ const StatusLabel = ({ status, text, customStyles }) => {
 };
 
 StatusLabel.propTypes = {
-  status: PropTypes.oneOf(["up", "down", "unknown", "cannot resolve"]),
+  status: PropTypes.oneOf(["up", "down", "pending", "cannot resolve"]),
   text: PropTypes.string,
   customStyles: PropTypes.object,
 };
